push_dir rejects conflicting directories
While a loader may have nested root directories to support
concerns in Rails, root directories among loaders must be
clearly separated.

This uses string comparison. Therefore, if a loader configures a root
directory that with symlinks that resolve to a certain realpath, and
another loader configures that same real path as root directory, the
current implementation does not detect the conflict.

I might address that in the future.