RBS notation for methods and attributes
We have now a standard notation to write signatures, let's adopt it.

This stays at the annotation level only, uses `@sig` as a custom tag for visual
coherence with `@private` and others, and does not define interface types, but
keeps duck typing notation as in

    #call | #debug | nil

which is not RBS.

It is a first step. Perhaps we integrate RBS properly in the future. In any
case, I believe we'd have *.rbs files generated, because I want the signatures
right there in the source code to help reading it. Time will tell.