Remove all calls to realpath
It is unclear that we have to work with the real path of root dirs.

Users are expected to be consistent in the kind of paths they use, so ignored
and collapsed stuff could be mistaken, and it is dubious that we need to go to
the trouble of mix and match them, specially with globs.

Also, it is good usability to see the same paths returned by the API and seen in
the logs.

In addition, require_dependency in Rails searches for an absolute path
concatenating whatever is in autoload_paths with the given file.

Real paths affect the support for require using a relative file name, which is
an experimental and undocumented feature, so we have some margin.

It might be the case that we revisit this, but by now, let's keep it simple.