Extract loader helpers to a separate module
Next step in trimming lib/zeitwerk/loader.rb is to move auxiliary methods out.

With this, we reach a sweet spot in which the core loader implementation is what
remains in the file. This is interesting, because the bulk of the public API is
configuration, but that's in a mixin. From the point of view of the library, the
main code is the one that does the actual work, everything else is auxiliary.
That's the code I want to see when I open lib/zeitwerk/loader.rb.