Restore support for namespaces that are not hashable
2.5.0 introduced a hash table whose keys were pairs (parent, cname), in order to
be able to say if a given pair had an autoload set by us.

That broke a project that has a namespace with the `hash` method overridden with
a different arity. Hence, not hashable.

This patch restores the structure used in 2.4, and implements a different way to
answer that question. Indeed, this alternative is simpler.

Fixes #188.