Make on_unload callbacks more robust in the event of failed autoloads
The state of the internal metadata after failed autoloads varies with Ruby
versions. There was a change in 3.1, and right now in `head` const_defined?
returns still true if the expected constant was not defined.

So, the cdef? that was guarding the execution of on_unload callbacks started to
fail in the event of failed autoloads on Ruby `head`.

It has to be seen whether this is considered to be a regression, but in any
case, the way Ruby behaves on failed autoloads is not specified as of this
writing, so an optimistic const_get seems more robust.