Unlock the reloading mutex in an ensure clause
The practical utility of this is unclear. If the wrapped calls raise, the loader
is left in an inconsistent state and is surely unusable anyway. This is why in a
first implementation this clause was not written.

However, even in that situation, the fact is that you are no longer reloading.
So, ensuring the reloading state is cleared seems just right.