Fixes portability of tmp directory removal in the test suite
The test suite started failing on `windows-latest/head`.

The second removal of the temporary directory used to generate test projects was
not being successful on Windows using the `main` branch of Ruby. Problem was, we
were within that directory by then, and Windows complains the directory is in
use. But `Errno::EACCES` was being eaten by `FileUtils.rm_rf`.

The helper where this happens starts recreating that directory to ensure a clean
slate, so the suite still worked correctly. The purpose of the second removal
was to not leave garbage behind, which is not critical for the suite to pass.

I believe this is the commit that revealed this flaw:

    FileUtils.rm* methods swallows only Errno::ENOENT when force is true
    https://github.com/ruby/ruby/commit/983115cf3c8f75b1afbe3274f02c1529e1ce3a81