Let eager_load_dir not raise on ignored arguments
This is an edge case.

Context, we allow root directories to be ignored, not an error condition.
Similary, recursive eager loading skips ignored descendants. In general, the
behaviour is that ignoring has higher priority, when in conflict, it wins.

So, to me, it is more consistent to not err if the argument is already ignored
in the first place.